# Fix data race when recording metrics and spans