Resolve #496. This is a corner case where a data race would happen.

The performance is also improved for most cases:

```
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/XSAM/otelsql
cpu: Apple M1 Max
                                              │   old.txt   │               new.txt               │
                                              │   sec/op    │   sec/op     vs base                │
RecordMetric/InstrumentAttributesGetter/5-10    609.8n ± 3%   443.6n ± 2%  -27.26% (p=0.000 n=10)
RecordMetric/InstrumentAttributesGetter/10-10   1.471µ ± 1%   1.347µ ± 2%   -8.46% (p=0.000 n=10)
CreateSpan/AttributesGetter/5-10                287.9n ± 6%   277.7n ± 1%   -3.54% (p=0.000 n=10)
CreateSpan/AttributesGetter/10-10               387.9n ± 1%   525.5n ± 1%  +35.46% (p=0.000 n=10)
geomean                                         562.6n        543.3n        -3.42%

                                              │   old.txt    │                new.txt                 │
                                              │     B/op     │     B/op      vs base                  │
RecordMetric/InstrumentAttributesGetter/5-10    1.781Ki ± 0%   1.219Ki ± 0%  -31.58% (p=0.000 n=10)
RecordMetric/InstrumentAttributesGetter/10-10   4.344Ki ± 0%   3.719Ki ± 0%  -14.39% (p=0.000 n=10)
CreateSpan/AttributesGetter/5-10                  824.0 ± 0%     824.0 ± 0%        ~ (p=1.000 n=10) ¹
CreateSpan/AttributesGetter/10-10               1.180Ki ± 0%   1.805Ki ± 0%  +52.98% (p=0.000 n=10)
geomean                                         1.646Ki        1.602Ki        -2.71%
¹ all samples are equal

                                              │  old.txt   │               new.txt                │
                                              │ allocs/op  │ allocs/op   vs base                  │
RecordMetric/InstrumentAttributesGetter/5-10    6.000 ± 0%   5.000 ± 0%  -16.67% (p=0.000 n=10)
RecordMetric/InstrumentAttributesGetter/10-10   7.000 ± 0%   7.000 ± 0%        ~ (p=1.000 n=10) ¹
CreateSpan/AttributesGetter/5-10                9.000 ± 0%   8.000 ± 0%  -11.11% (p=0.000 n=10)
CreateSpan/AttributesGetter/10-10               9.000 ± 0%   9.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                         7.637        7.085        -7.23%
¹ all samples are equal
```

This performance can be further improved, but it is not within the scope of this PR.